slider wiper with zoom OK, files Zoomable-img-osd and Slider-osd for Openseadragon in repo and functionnal
