+++ src/pages/weaving/[slug].astro
@@ -2,6 +2,7 @@
 import fetchApi from "../../lib/strapi";
 import Layout from "../../layouts/Layout.astro";
 import Slider from '../../components/slider/Slider.jsx';
+// import SliderOsd from '../../components/slider/Slider-osd.jsx';
 
 
 export async function getStaticPaths() {
@@ -42,7 +43,9 @@ export const prerender = true;
     <h1 class="font-bradford text-base fixed top-[20px] left-[20px] z-10">
         {weaving?.title || "Titre non disponible"}
     </h1>
-    
+
     <Slider client:only medias={weaving?.medias || []} />
+    
+    <!-- <SliderOsd client:only medias={weaving?.medias || []} /> -->
     
 </Layout>
